requirements.txt
pydantic==2.8.2

google-api-python-client==2.137.0
google-auth==2.33.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

python-dotenv==1.0.1
rich==13.7.1
httpx==0.27.2

fastapi
uvicorn
google-genai